[P] handle exceptions when CS disconnects (broken pipe & empty msg)
When the comm server breaks the connection during Communicator's blocking
read operation, the Communicator reads an empty message. In that case the
`OperationCanceledException` is now thrown manually and handled in main
program method.

The other type of handled errors are situations, when the connection is
broken in the middle of Communicator's write operation. We handle the
`IOException` in main program method.

The handlers above are repeated twice: separately for listing games
and separately when starting the actual game.